Move modules files to mirror new pages structure
