+++ src/modules/_[spaceId]/_c/ChatPage/ChatPageNavbarExtension.tsx
